refactor splitter analysis code so as not to import from PYMEAcquire
