+++ PYME/Analysis/tile_pyramid.py
@@ -720,10 +720,10 @@ def tile_pyramid(out_folder, ds, xm, ym, mdh, split=False, skipMoveFrames=False,
     frameSizeX, frameSizeY, numFrames = ds.shape[:3]
     
     if split:
-        from PYME.Acquire.Hardware import splitter
+        from PYME.Analysis import splitting
         frameSizeY /= 2
         nchans = 2
-        unmux = splitter.Unmixer(shiftfield, mdh.voxelsize_nm.x)
+        unmux = splitting.Unmixer(shiftfield, mdh.voxelsize_nm.x)
     else:
         nchans = 1
