+++ PYME/DSView/modules/unmixing.py
@@ -23,8 +23,6 @@
 
 
 import wx
-#from PYME.Acquire.Hardware import splitter
-#from PYME.IO.DataSources import UnsplitDataSource
 import numpy as np
 
 #from PYME.DSView.arrayViewPanel import *
